requirements.txt
Django~=5.2.3
django-debug-toolbar~=5.2.0
psycopg2-binary~=2.9.10
python-dotenv~=1.1.1
requests~=2.32.4
celery~=5.5.3